Clears procedure data after adding procedure
Although it now seems to be clearing the data correctly, it crashes
after saving the data. At first, this was because the data was being
cleared before we had finished using it. This was fixed however, and now
there's another issue that needs to be found. At least the data is being
saved successfully.