Fixes ProcedureListFragment.ListProcedures method
After some debugging it turns out that the consumer method was being called
outside a Firestore listener, so it was always empty. Moving the consumer
call inside the listener fixed that problem. However, search results have a
title 'Search results for null', so the search term isn't being picked up
correctly. It's being picked up by the ProcedureListFragment correctly (i.e.
the search results are correct, it's just the title that's broken).