Adds new layouts for viewing a Procedure
Portrait and landscape views now have different layouts. Both use golden
proportions to separate steps from comments, and add a new background colour
behind the comments.

Future versions will probably change this layout. The comments are a bit
distracting, so perhaps they would be better in a navigation drawer or dialog.
Also the comment button looks a bit strange where it currently is. Perhaps
putting it as a menu item might be better.